--- Assets/Scripts/ColonyExecutionAuthor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 443419cc594441fdb3898347e9aee3c3
-timeCreated: 1648298117
