+++ Assets/Scripts/Data/Execution/ExecutionState.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Unity.Entities;
+using UnityEngine;
+
+[GenerateAuthoringComponent]
+public struct ExecutionState : IComponentData
+{
+    public short id;
+    public short executionLine;
+}
