+++ Assets/Scripts/Data/FoodLeft.cs
@@ -5,4 +5,9 @@ public struct FoodLeft : IComponentData
 {
     public int Left;
     public int Total;
+    public FoodLeft(int total, int left)
+    {
+        Total = total;
+        Left = left;
+    }
 }
