+++ Assets/Scripts/Data/GlobalData.cs
@@ -1,0 +1,10 @@
+﻿using Unity.Entities;
+
+namespace Data
+{
+    [GenerateAuthoringComponent]
+    public struct GlobalData : IComponentData
+    {
+        public float SpawnInterval;
+    }
+}
