+++ Assets/Scripts/Data/GlobalData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1d75afd5e039497da59d0170ccb0120f
+timeCreated: 1648431426
