+++ Assets/Scripts/Data/HybridStoreAuthor.cs
